refactor: move CQ SDK E2E test to packages/api +
- move CQ SDK E2E test to packages/api
- remove uneeded test code from CQ SDK

Ref. metriport/metriport-internal#2553
