+++ packages/carequality-sdk/src/index.ts
@@ -1,5 +1,6 @@
 export {
   APIMode,
+  CarequalityManagementApi,
   CarequalityManagementApi as CarequalityManagementAPI,
 } from "./client/carequality";
 export { CarequalityManagementApiFhir } from "./client/carequality-fhir";
